Actualizo la tabla sql

+++ backend/src/main/resources/db/migration/V1__initial_schema.sql
@@ -1,20 +1,20 @@
 -- Estructura de Tablas Generales
 CREATE TABLE barrios (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     nombre VARCHAR(100) NOT NULL,
     direccion VARCHAR(150) NOT NULL
 );
 
 CREATE TABLE unidades_funcionales (
-    id SERIAL PRIMARY KEY,
-    barrio_id INT REFERENCES barrios(id),
+    id BIGSERIAL PRIMARY KEY,
+    barrio_id BIGINT REFERENCES barrios(id),
     identificador VARCHAR(20) NOT NULL,
     tipo_unidad VARCHAR(20) NOT NULL
 );
 
 -- Tabla Padre para Herencia de Personas
 CREATE TABLE personas (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     nombre VARCHAR(100) NOT NULL,
     apellido VARCHAR(100) NOT NULL,
     dni VARCHAR(20) UNIQUE NOT NULL,
@@ -24,36 +24,36 @@ CREATE TABLE personas (
 
 -- Tablas Hijas para Herencia JOINED
 CREATE TABLE residentes (
-    id INT PRIMARY KEY REFERENCES personas(id) ON DELETE CASCADE,
-    unidad_funcional_id INT REFERENCES unidades_funcionales(id)
+    id BIGINT PRIMARY KEY REFERENCES personas(id) ON DELETE CASCADE,
+    unidad_funcional_id BIGINT REFERENCES unidades_funcionales(id)
 );
 
 CREATE TABLE proveedores (
-    id INT PRIMARY KEY REFERENCES personas(id) ON DELETE CASCADE,
+    id BIGINT PRIMARY KEY REFERENCES personas(id) ON DELETE CASCADE,
     tipo_servicio VARCHAR(50) NOT NULL
 );
 
 CREATE TABLE personal_seguridad (
-    id INT PRIMARY KEY REFERENCES personas(id) ON DELETE CASCADE
+    id BIGINT PRIMARY KEY REFERENCES personas(id) ON DELETE CASCADE
 );
 
 CREATE TABLE personal_mantenimiento (
-    id INT PRIMARY KEY REFERENCES personas(id) ON DELETE CASCADE
+    id BIGINT PRIMARY KEY REFERENCES personas(id) ON DELETE CASCADE
 );
 
 CREATE TABLE visitantes (
-    id INT PRIMARY KEY REFERENCES personas(id) ON DELETE CASCADE
+    id BIGINT PRIMARY KEY REFERENCES personas(id) ON DELETE CASCADE
 );
 
 CREATE TABLE administradores (
-    id INT PRIMARY KEY REFERENCES personas(id) ON DELETE CASCADE
+    id BIGINT PRIMARY KEY REFERENCES personas(id) ON DELETE CASCADE
 );
 
 -- Tabla de Autorizaciones de Ingreso (creadas por Residentes)
 CREATE TABLE autorizaciones_ingreso (
-    id SERIAL PRIMARY KEY,
-    residente_id INT REFERENCES residentes(id),
-    visitante_id INT REFERENCES visitantes(id),
+    id BIGSERIAL PRIMARY KEY,
+    residente_id BIGINT REFERENCES residentes(id),
+    visitante_id BIGINT REFERENCES visitantes(id),
     fecha_desde TIMESTAMP NOT NULL,
     fecha_hasta TIMESTAMP NOT NULL,
     utilizada BOOLEAN DEFAULT FALSE
@@ -61,40 +61,40 @@ CREATE TABLE autorizaciones_ingreso (
 
 -- Tabla de Visitas / Log de Entrada y Salida
 CREATE TABLE visitas (
-    id SERIAL PRIMARY KEY,
-    visitante_id INT REFERENCES visitantes(id),
-    autorizacion_id INT REFERENCES autorizaciones_ingreso(id),
+    id BIGSERIAL PRIMARY KEY,
+    visitante_id BIGINT REFERENCES visitantes(id),
+    autorizacion_id BIGINT REFERENCES autorizaciones_ingreso(id),
     fecha_ingreso TIMESTAMP NOT NULL,
     fecha_salida TIMESTAMP,
     estado VARCHAR(20) NOT NULL,
-    registrado_por_seguridad_id INT REFERENCES personal_seguridad(id)
+    registrado_por_seguridad_id BIGINT REFERENCES personal_seguridad(id)
 );
 
 -- Tabla de Reclamos / Incidentes
 CREATE TABLE reclamos (
-    id SERIAL PRIMARY KEY,
-    residente_id INT REFERENCES residentes(id),
+    id BIGSERIAL PRIMARY KEY,
+    residente_id BIGINT REFERENCES residentes(id),
     tipo_reclamo VARCHAR(30) NOT NULL,
     descripcion TEXT NOT NULL,
     prioridad VARCHAR(20) NOT NULL,
     estado VARCHAR(20) NOT NULL,
     fecha_creacion TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
-    responsable_id INT REFERENCES personas(id)
+    responsable_id BIGINT REFERENCES personas(id)
 );
 
 -- Tabla de Tareas de Mantenimiento derivadas de un Reclamo o creadas por Admin
 CREATE TABLE tareas_mantenimiento (
-    id SERIAL PRIMARY KEY,
-    reclamo_id INT REFERENCES reclamos(id) ON DELETE SET NULL,
+    id BIGSERIAL PRIMARY KEY,
+    reclamo_id BIGINT REFERENCES reclamos(id) ON DELETE SET NULL,
     descripcion TEXT NOT NULL,
     estado VARCHAR(20) NOT NULL,
-    responsable_id INT REFERENCES personas(id)
+    responsable_id BIGINT REFERENCES personas(id)
 );
 
 -- Historial de Estado de Reclamos
 CREATE TABLE historial_estado_reclamos (
-    id SERIAL PRIMARY KEY,
-    reclamo_id INT REFERENCES reclamos(id) ON DELETE CASCADE,
+    id BIGSERIAL PRIMARY KEY,
+    reclamo_id BIGINT REFERENCES reclamos(id) ON DELETE CASCADE,
     estado_anterior VARCHAR(20),
     estado_nuevo VARCHAR(20) NOT NULL,
     fecha_cambio TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
